Use short for texpatch_t .originx, .originy: fixes problem caused by sign conversion in the endianness code.
Subversion-branch: /trunk/chocolate-doom
Subversion-revision: 961

+++ src/r_data.c
@@ -98,8 +98,8 @@ typedef struct
     // Block origin (allways UL),
     // which has allready accounted
     // for the internal origin of the patch.
-    int		originx;	
-    int		originy;
+    short	originx;	
+    short	originy;
     int		patch;
 } texpatch_t;
